requirements.txt
openai>=1.0.0
fal-client>=0.13.0
requests>=2.31.0
python-dotenv>=1.0.0
